convert to maven project
