+++ .gitignore
@@ -37,3 +37,4 @@ local.properties
 # ignore compiled javadoc
 
 doc/javadoc/
+/target/
